refactor stories.js by moving validation logic into the utils.js and fix a bug in the update of a story
