+++ backend/routes/stories.js
@@ -1,11 +1,16 @@
 const express = require("express");
-const { check } = require("express-validator");
-const { handleValidationErrors, asyncHandler } = require("../utils");
+
 const { requireAuth } = require("../auth");
 const router = express.Router();
 const db = require("../db/models");
 
 const { Story, User } = db;
+
+const {
+  handleValidationErrors,
+  asyncHandler,
+  validateStory,
+} = require("../utils");
 
 router.use(requireAuth);
 
@@ -46,25 +51,11 @@ router.get(
   })
 );
 
-const validateStory = [
-  check("title")
-    .exists({ checkFalsy: true })
-    .withMessage("Story title can't be empty."),
-  check("body")
-    .exists({ checkFalsy: true })
-    .withMessage("Story body can't be empty."),
-  //  message cannot be longer than 280 characters:
-  check("byline")
-    .isLength({ max: 140 })
-    .withMessage("Byline can't be longer than140 characters."),
-  handleValidationErrors,
-  //TODO add the remaining validation
-];
-
 //Create a new sotry
 router.post(
   "/",
   validateStory,
+  handleValidationErrors,
   asyncHandler(async (req, res) => {
     const userId = "1"; //TODO replace with  this once it is connected with the front end ====> req.user.id;
     const { title, byline, body, image } = req.body; //TODO anything else?
@@ -84,12 +75,14 @@ router.post(
 router.put(
   "/:id",
   validateStory,
+  handleValidationErrors,
   asyncHandler(async (req, res, next) => {
-    const userId = "1"; //TODO replace with  this once it is connected with the front end ====> req.user.id;
+    const userId = "1"; // req.user.id; //TODO replace with  this once it is connected with the front end ====> req.user.id;
+    const storyId = req.params.id;
     const story = await Story.findOne({
-      where: { id: userId },
+      where: { id: storyId },
     });
-    console.log("=====", story.userId, userId);
+
     if (userId.toString() !== story.userId.toString()) {
       const err = new Error("Unauthorized");
       err.status = 401;
@@ -97,12 +90,13 @@ router.put(
       err.title = "Unauthorized";
       throw err;
     }
+
     if (story) {
       const { title, byline, body, image } = req.body;
       await story.update({ title, byline, body, image }); //TODO should we pass in the userId as well?
       res.json({ story });
     } else {
-      next(storyNotFoundError(req.params.id));
+      next(storyNotFoundError(storyId));
     }
   })
 );
